requirements.txt
dnspython==2.7.0
requests==2.32.3
socks==0
